# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stock {
  ticker: string;
  name: string;
  currentPrice: number;
  change: number;
  changePercent: number;
  volume: number;
  marketCap: number;
  sector: string;
}

export interface StockDetail {
  ticker: string;
  name: string;
  currentPrice: number;
  change: number;
  changePercent: number;
  volume: number;
  marketCap: number;
  sector: string;
  high52w: number;
  low52w: number;
  per: number;
  pbr: number;
  eps: number;
  dividendYield: number;
  openPrice: number;
  highPrice: number;
  lowPrice: number;
}

export interface PricePoint {
  date: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface NewsItem {
  id: string;
  title: string;
  summary: string;
  source: string;
  publishedAt: string;
  url: string;
}

export interface IndexData {
  name: string;
  value: number;
  change: number;
  changePercent: number;
}

export interface MarketSummary {
  kospi: IndexData;
  kosdaq: IndexData;
  usdKrw: IndexData;
  sp500: IndexData;
  nasdaq: IndexData;
  updatedAt: string;
}

export type CreateUserRequestAvatar =
  (typeof CreateUserRequestAvatar)[keyof typeof CreateUserRequestAvatar];

export const CreateUserRequestAvatar = {
  male: "male",
  female: "female",
} as const;

export type CreateUserRequestDifficulty =
  (typeof CreateUserRequestDifficulty)[keyof typeof CreateUserRequestDifficulty];

export const CreateUserRequestDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  expert: "expert",
} as const;

export interface CreateUserRequest {
  username: string;
  avatar: CreateUserRequestAvatar;
  difficulty: CreateUserRequestDifficulty;
}

export interface User {
  id: number;
  username: string;
  avatar: string;
  difficulty: string;
  seedMoney: number;
  cashBalance: number;
  totalAssets: number;
  totalReturn: number;
  totalReturnPercent: number;
  createdAt: string;
}

export interface Holding {
  ticker: string;
  name: string;
  shares: number;
  avgPrice: number;
  currentPrice: number;
  evaluationAmount: number;
  profitLoss: number;
  profitLossPercent: number;
}

export interface Portfolio {
  userId: number;
  cashBalance: number;
  totalAssets: number;
  totalReturn: number;
  totalReturnPercent: number;
  holdings: Holding[];
}

export type TradeType = (typeof TradeType)[keyof typeof TradeType];

export const TradeType = {
  buy: "buy",
  sell: "sell",
} as const;

export interface Trade {
  id: number;
  userId: number;
  ticker: string;
  stockName: string;
  type: TradeType;
  shares: number;
  price: number;
  totalAmount: number;
  createdAt: string;
}

export type ExecuteTradeRequestType =
  (typeof ExecuteTradeRequestType)[keyof typeof ExecuteTradeRequestType];

export const ExecuteTradeRequestType = {
  buy: "buy",
  sell: "sell",
} as const;

export interface ExecuteTradeRequest {
  userId: number;
  ticker: string;
  type: ExecuteTradeRequestType;
  shares: number;
}

export interface TradeResult {
  success: boolean;
  trade: Trade;
  newCashBalance: number;
  message: string;
}

export interface TradeError {
  success: boolean;
  message: string;
}

export interface RankingEntry {
  rank: number;
  userId: number;
  username: string;
  avatar: string;
  difficulty: string;
  seedMoney: number;
  totalAssets: number;
  totalReturn: number;
  totalReturnPercent: number;
}

export type GetStockHistoryParams = {
  period?: GetStockHistoryPeriod;
};

export type GetStockHistoryPeriod =
  (typeof GetStockHistoryPeriod)[keyof typeof GetStockHistoryPeriod];

export const GetStockHistoryPeriod = {
  "1d": "1d",
  "1w": "1w",
  "1m": "1m",
  "3m": "3m",
  "1y": "1y",
} as const;

export type GetRankingsParams = {
  difficulty?: GetRankingsDifficulty;
};

export type GetRankingsDifficulty =
  (typeof GetRankingsDifficulty)[keyof typeof GetRankingsDifficulty];

export const GetRankingsDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  expert: "expert",
  all: "all",
} as const;