# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Toss Securities Stock Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stocks
    description: Stock operations
  - name: users
    description: User operations
  - name: trades
    description: Trade operations
  - name: rankings
    description: Ranking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stocks:
    get:
      operationId: getStocks
      tags: [stocks]
      summary: Get all stocks
      responses:
        "200":
          description: List of stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stock"
  /stocks/{ticker}:
    get:
      operationId: getStockByTicker
      tags: [stocks]
      summary: Get stock by ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockDetail"
        "404":
          description: Stock not found
  /stocks/{ticker}/history:
    get:
      operationId: getStockHistory
      tags: [stocks]
      summary: Get stock price history
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, 1y]
            default: 1m
      responses:
        "200":
          description: Stock price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
  /stocks/{ticker}/news:
    get:
      operationId: getStockNews
      tags: [stocks]
      summary: Get stock related news
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
  /market/summary:
    get:
      operationId: getMarketSummary
      tags: [stocks]
      summary: Get market summary
      responses:
        "200":
          description: Market summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketSummary"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
  /users/{userId}/portfolio:
    get:
      operationId: getUserPortfolio
      tags: [users]
      summary: Get user portfolio
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /users/{userId}/trades:
    get:
      operationId: getUserTrades
      tags: [users]
      summary: Get user trade history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /trades:
    post:
      operationId: executeTrade
      tags: [trades]
      summary: Execute a buy or sell trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteTradeRequest"
      responses:
        "200":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Trade failed (insufficient funds/shares)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeError"
  /rankings:
    get:
      operationId: getRankings
      tags: [rankings]
      summary: Get rankings by difficulty
      parameters:
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, expert, all]
            default: all
      responses:
        "200":
          description: Rankings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Stock:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        currentPrice:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        marketCap:
          type: number
        sector:
          type: string
      required:
        - ticker
        - name
        - currentPrice
        - change
        - changePercent
        - volume
        - marketCap
        - sector
    StockDetail:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        currentPrice:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        marketCap:
          type: number
        sector:
          type: string
        high52w:
          type: number
        low52w:
          type: number
        per:
          type: number
        pbr:
          type: number
        eps:
          type: number
        dividendYield:
          type: number
        openPrice:
          type: number
        highPrice:
          type: number
        lowPrice:
          type: number
      required:
        - ticker
        - name
        - currentPrice
        - change
        - changePercent
        - volume
        - marketCap
        - sector
        - high52w
        - low52w
        - per
        - pbr
        - eps
        - dividendYield
        - openPrice
        - highPrice
        - lowPrice
    PricePoint:
      type: object
      properties:
        date:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required:
        - date
        - open
        - high
        - low
        - close
        - volume
    NewsItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        url:
          type: string
      required:
        - id
        - title
        - summary
        - source
        - publishedAt
        - url
    MarketSummary:
      type: object
      properties:
        kospi:
          $ref: "#/components/schemas/IndexData"
        kosdaq:
          $ref: "#/components/schemas/IndexData"
        usdKrw:
          $ref: "#/components/schemas/IndexData"
        sp500:
          $ref: "#/components/schemas/IndexData"
        nasdaq:
          $ref: "#/components/schemas/IndexData"
        updatedAt:
          type: string
      required:
        - kospi
        - kosdaq
        - usdKrw
        - sp500
        - nasdaq
        - updatedAt
    IndexData:
      type: object
      properties:
        name:
          type: string
        value:
          type: number
        change:
          type: number
        changePercent:
          type: number
      required:
        - name
        - value
        - change
        - changePercent
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        avatar:
          type: string
          enum: [male, female]
        difficulty:
          type: string
          enum: [beginner, intermediate, expert]
      required:
        - username
        - avatar
        - difficulty
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: string
        difficulty:
          type: string
        seedMoney:
          type: number
        cashBalance:
          type: number
        totalAssets:
          type: number
        totalReturn:
          type: number
        totalReturnPercent:
          type: number
        createdAt:
          type: string
      required:
        - id
        - username
        - avatar
        - difficulty
        - seedMoney
        - cashBalance
        - totalAssets
        - totalReturn
        - totalReturnPercent
        - createdAt
    Portfolio:
      type: object
      properties:
        userId:
          type: integer
        cashBalance:
          type: number
        totalAssets:
          type: number
        totalReturn:
          type: number
        totalReturnPercent:
          type: number
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/Holding"
      required:
        - userId
        - cashBalance
        - totalAssets
        - totalReturn
        - totalReturnPercent
        - holdings
    Holding:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        shares:
          type: number
        avgPrice:
          type: number
        currentPrice:
          type: number
        evaluationAmount:
          type: number
        profitLoss:
          type: number
        profitLossPercent:
          type: number
      required:
        - ticker
        - name
        - shares
        - avgPrice
        - currentPrice
        - evaluationAmount
        - profitLoss
        - profitLossPercent
    Trade:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        ticker:
          type: string
        stockName:
          type: string
        type:
          type: string
          enum: [buy, sell]
        shares:
          type: number
        price:
          type: number
        totalAmount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - userId
        - ticker
        - stockName
        - type
        - shares
        - price
        - totalAmount
        - createdAt
    ExecuteTradeRequest:
      type: object
      properties:
        userId:
          type: integer
        ticker:
          type: string
        type:
          type: string
          enum: [buy, sell]
        shares:
          type: number
      required:
        - userId
        - ticker
        - type
        - shares
    TradeResult:
      type: object
      properties:
        success:
          type: boolean
        trade:
          $ref: "#/components/schemas/Trade"
        newCashBalance:
          type: number
        message:
          type: string
      required:
        - success
        - trade
        - newCashBalance
        - message
    TradeError:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    RankingEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatar:
          type: string
        difficulty:
          type: string
        seedMoney:
          type: number
        totalAssets:
          type: number
        totalReturn:
          type: number
        totalReturnPercent:
          type: number
      required:
        - rank
        - userId
        - username
        - avatar
        - difficulty
        - seedMoney
        - totalAssets
        - totalReturn
        - totalReturnPercent