# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserRequestAvatar } from "./createUserRequestAvatar";
import type { CreateUserRequestDifficulty } from "./createUserRequestDifficulty";

export interface CreateUserRequest {
  username: string;
  avatar: CreateUserRequestAvatar;
  difficulty: CreateUserRequestDifficulty;
}