# lib/api-zod/src/generated/types/executeTradeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExecuteTradeRequestType } from "./executeTradeRequestType";

export interface ExecuteTradeRequest {
  userId: number;
  ticker: string;
  type: ExecuteTradeRequestType;
  shares: number;
}