# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Holding {
  ticker: string;
  name: string;
  shares: number;
  avgPrice: number;
  currentPrice: number;
  evaluationAmount: number;
  profitLoss: number;
  profitLossPercent: number;
}