# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: string;
  title: string;
  summary: string;
  source: string;
  publishedAt: string;
  url: string;
}