# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toss Securities Stock Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Holding } from "./holding";

export interface Portfolio {
  userId: number;
  cashBalance: number;
  totalAssets: number;
  totalReturn: number;
  totalReturnPercent: number;
  holdings: Holding[];
}